refactor(chord-keyboard): remove preact for simpler SVG rendering
