+++ extensions/pianoman/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Fix] - {PR_MERGE_DATE}
 
-- Fix: Resolve recurring "Incompatible React versions: react and react-dom must have the exact same version" crash by removing `react-dom`. Chord keyboard SVGs are now rendered with Preact (`preact-render-to-string`) in an isolated module, so string rendering no longer depends on the host React version.
+- Fix: Resolve recurring "Incompatible React versions: react and react-dom must have the exact same version" crash by removing `react-dom`. Chord keyboard SVGs are now built as plain strings (no server-side render library), so string rendering no longer depends on the host React version.
 
 ## [Fix] - 2025-04-16
